+++ R/RowLinkedMatrix.R
@@ -1,76 +1,84 @@
-subset.RowLinkedMatrix <- function(x, i, j, ..., drop = TRUE) {
-    # Check indices and dimensions
-    nX <- nrow(x)
-    pX <- ncol(x)
-    if (missing(i)) {
-        i <- 1L:nX
-    }
-    if (missing(j)) {
-        j <- 1L:pX
-    }
-    if (typeof(i) == "logical") {
-        i <- rep_len(i, nX)
-        i <- which(i)
-    } else if (typeof(i) == "character") {
-        i <- match(i, rownames(x))
-    } else if (typeof(i) == "double") {
-        i <- as.integer(i)
-    }
-    if (typeof(j) == "logical") {
-        j <- rep_len(j, pX)
-        j <- which(j)
-    } else if (typeof(j) == "character") {
-        j <- match(j, colnames(x))
-    } else if (typeof(j) == "double") {
-        j <- as.integer(j)
-    }
-    n <- length(i)
-    p <- length(j)
-    if (p > pX || n > nX) {
-        stop("Either the number of columns or number of rows requested exceed the number of rows or columns in x, try dim(x)...")
-    }
-    # Providing a sorted row index will eliminate the need to reorder the
-    # result matrix later (avoiding a copy)
-    isUnsorted <- is.unsorted(i)
-    if (isUnsorted) {
-        # Reorder rows for sequential retrieval by node
-        originalOrder <- rank(i, ties.method = "first")
-        sortedRows <- sort(i)
+extract_matrix.RowLinkedMatrix <- function(x, i, j, ...) {
+    # Handle x[FALSE, ]
+    if (length(i) == 0L) {
+        Z <- matrix(integer(), nrow = 0L, ncol = length(j), dimnames = list(NULL, colnames(x)[j]))
     } else {
-        sortedRows <- i
-    }
-    # Compute node inventory
-    globalIndex <- index(x, i = sortedRows, sort = FALSE)
-    whichNodes <- unique(globalIndex[, 1L])
-    # If there are several nodes involved, aggregate the result in a separate
-    # matrix, otherwise pass through result
-    if (length(whichNodes) > 1L) {
-        # Initialize result matrix as integer matrix because it does not take up as
-        # much space as double() but is more useful than logical()
-        Z <- matrix(data = integer(), nrow = n, ncol = p)
-        # Use dimnames instead of rownames and colnames to avoid copy
-        dimnames(Z) <- list(rownames(x)[sortedRows], colnames(x)[j])
-        end <- 0L
-        for (k in whichNodes) {
-            localIndex <- globalIndex[globalIndex[, 1L] == k, , drop = FALSE]
-            ini <- end + 1L
-            end <- ini + nrow(localIndex) - 1L
-            # Convert to matrix to support data frames
-            Z[ini:end, ] <- as.matrix(x[[k]][localIndex[, 3L], j, drop = FALSE])
+        # Determine nodes and node boundaries for query
+        index <- index(x, i = i, sort = FALSE)
+        nodeList <- unique(index[, 1L])
+        # If there are several nodes involved in the query, aggregate the
+        # result in a separate matrix, otherwise pass through result
+        if (length(nodeList) > 1L) {
+            # Initialize result matrix as integer() because it does not take up
+            # as much space as double() but is more common than logical()
+            Z <- matrix(data = integer(), nrow = length(i), ncol = length(j), dimnames = list(rownames(x)[i], colnames(x)[j]))
+            for (curNode in nodeList) {
+                if (is.na(curNode)) {
+                    nodeIndex = is.na(index[, 1L])
+                    Z[nodeIndex, ] <- NA_integer_
+                } else {
+                    nodeIndex <- index[, 1L] == curNode
+                    nodeIndex[is.na(nodeIndex)] <- FALSE
+                    # Convert to matrix to support data frames
+                    Z[nodeIndex, ] <- as.matrix(x[[curNode]][index[nodeIndex, 3L], j, drop = FALSE])
+                }
+            }
+        } else {
+            # Handle x[NA, ]
+            if (is.na(nodeList)) {
+                Z <- matrix(NA_integer_, nrow = length(i), ncol = length(j), dimnames = list(rep(NA_character_, length(i)), colnames(x)[j]))
+            } else {
+                # Convert to matrix to support data frames
+                Z <- as.matrix(x[[nodeList]][index[, 3L], j, drop = FALSE])
+            }
         }
+    }
+    return(Z)
+}
+
+
+extract_vector.RowLinkedMatrix <- function(x, i, ...) {
+    if (length(i) == 0L) {
+        Z <- integer(0L)
     } else {
-        Z <- as.matrix(x[[whichNodes]][globalIndex[, 3L], j, drop = FALSE])
+        # Convert one-dimensional index to two-dimensional index
+        ij <- crochet:::ktoij(x, i)
+        # Determine nodes and node boundaries for query (in this case we cannot
+        # use index() as rowsPerNode is needed to recalculate the single index)
+        rowsPerNode <- sapply(x, nrow)
+        nodeBoundaries <- c(0L, cumsum(rowsPerNode))
+        nodeMembership <- .bincode(ij$i, nodeBoundaries)
+        nodeList <- unique(nodeMembership)
+        # If there are several nodes involved in the query, aggregate the
+        # result in a separate matrix, otherwise pass through result
+        if (length(nodeList) > 1L) {
+            # Initialize result vector as integer() because it does not take up
+            # as much space as double() but is more common than logical()
+            Z <- vector(mode = "integer", length = length(i))
+            for (curNode in nodeList) {
+                if (is.na(curNode)) {
+                    nodeIndex <- is.na(nodeMembership)
+                    Z[nodeIndex] <- NA_integer_
+                } else {
+                    nodeIndex <- nodeMembership == curNode
+                    nodeIndex[is.na(nodeIndex)] <- FALSE
+                    # Convert two-dimensional index back to one-dimensional index
+                    localIndex <- ((ij$j[nodeIndex] - 1L) * rowsPerNode[curNode] + ij$i[nodeIndex]) - nodeBoundaries[curNode]
+                    Z[nodeIndex] <- x[[curNode]][localIndex]
+                }
+            }
+        } else {
+            # Handle x[NA]
+            if (is.na(nodeList)) {
+                Z <- rep(NA_integer_, length(i))
+            } else {
+                # Convert two-dimensional index back to one-dimensional index
+                localIndex <- ((ij$j - 1L) * rowsPerNode[nodeList] + ij$i) - nodeBoundaries[nodeList]
+                Z <- x[[nodeList]][localIndex]
+            }
+        }
     }
-    if (isUnsorted) {
-        # Return to original order
-        Z <- Z[originalOrder, , drop = FALSE]
-    }
-    if (drop == TRUE && (n == 1L || p == 1L)) {
-        # Let R handle drop behavior
-        return(Z[, ])
-    } else {
-        return(Z)
-    }
+    return(Z)
 }
 
 
@@ -232,11 +240,9 @@ index.RowLinkedMatrix <- function(x, i = NULL, sort = TRUE, ...) {
     } else {
         i <- seq_len(nodes[nrow(nodes), 3L])
     }
-    index <- matrix(data = integer(), nrow = length(i), ncol = 3L, dimnames = list(NULL, c("node", "row.global", "row.local")))
-    whichNode <- .bincode(i, breaks = c(0L, nodes[, 3L]))
-    index[, 1L] <- whichNode
-    index[, 2L] <- i
-    index[, 3L] <- i - nodes[whichNode, 2L] + 1L
+    nodeBoundaries <- c(0L, nodes[, 3L])
+    nodeMembership <- .bincode(i, breaks = nodeBoundaries)
+    index <- matrix(c(nodeMembership, i, i - nodeBoundaries[nodeMembership]), nrow = length(i), ncol = 3L, dimnames = list(NULL, c("node", "row.global", "row.local")))
     return(index)
 }
 
@@ -290,7 +296,7 @@ setMethod("initialize", signature(.Object = "RowLinkedMatrix"), function(.Object
 
 
 #' @export
-`[.RowLinkedMatrix` <- subset.RowLinkedMatrix
+`[.RowLinkedMatrix` <- crochet::extract(extract_vector = extract_vector.RowLinkedMatrix, extract_matrix = extract_matrix.RowLinkedMatrix)
 
 
 #' @export